Added put calls; refined seeded data; added debugging features

+++ migrations/20211116211414_lightningRule_table.js
@@ -3,13 +3,13 @@ exports.up = function(knex) {
     return knex.schema.createTable('lightning_rule', table => {
         table.increments('id');
         table.string('constraint_name').notNullable();
-        table.integer('constraint_parameter_integer');
+        table.bigInteger('constraint_parameter_integer');
         table.boolean('constraint_parameter_boolean');
-        table.integer('user_input_integer');
+        table.bigInteger('user_input_integer');
         table.boolean('user_input_boolean')
      })
   };
-  
+
   exports.down = function(knex) {
     return knex.schema.dropTableIfExists('lightning_rule')
   };
